requirements.txt
Flask==3.1.3
Pillow==12.2.0
pywebview>=4.0
pytest==9.0.3
pytest-cov==7.1.0
